% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{abs_to_col}
\alias{abs_to_col}
\title{Convert an absorbance to a hexidecimal color}
\usage{
abs_to_col(abs, pal)
}
\arguments{
\item{abs}{Numeric. Absorbances.}

\item{pal}{Character. A vector of hexidecimal colors.}
}
\value{
Character. Hexidecimal colors corresponding to absorbances.
}
\description{
Takes an absorbance and converts it to a hexidecimal color. For the default
qp_pal palette, this should provide a color that approximates real life
color at the given absorbance.
}
\details{
The absorbances have typical baseline absorbance (~ 0.07) removed,
and then an index is calculated with a logistic curve of maximum 100 and a
center of 0.15.
}
