% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.R
\name{qp}
\alias{qp}
\title{Quantify protein concentration from a MicroBCA assay}
\usage{
qp(
  x,
  replicate_orientation = c("h", "v"),
  sample_names = NULL,
  remove_empty = TRUE,
  ignore_outliers = c("all", "samples", "standards", "none"),
  standard_scale = c(0, 2^((2:7) - 5)),
  n_replicates = 3,
  wavelength = 562
)
}
\arguments{
\item{x}{A \code{spectramax}, \code{gp}, or \code{data.frame} object, or path to SPECTRAmax
.xls(x)/.txt file.}

\item{replicate_orientation}{Either 'h' or 'v' - see Details.}

\item{sample_names}{Character vector of sample names.}

\item{remove_empty}{Should wells that have less absorbance than the lowest
standard be dropped?}

\item{ignore_outliers}{Character. From which group - samples or standards - should
outliers be detected and removed?}

\item{standard_scale}{Numeric. Known concentrations of standards, in the order
they appear.}

\item{n_replicates}{Numeric. The number of techinical replicates.}

\item{wavelength}{Numeric. The wavelength absorbance was captured.}
}
\value{
a \code{tibble}
}
\description{
Quantify protein concentration from a MicroBCA assay
}
\details{
If \code{x} is a \code{spectramax}, the standards must start in the upper left
corner in the order dictated by \code{standard_scale}. Whether this is from from
left to right or top to bottom can be specified in \code{replicate_orientation}.
Note that \code{replicate_orientation} specified the direction that REPLICATES
lie, NOT the direction the samples flow (which will be perpendicular to
the replicates).

Note: \code{replicate_orientation}, \code{n_replicates}, and \code{wavelength} will
be silently ignored if \code{x} is not a \code{spectramax} or path to a
\code{spectramax}
}
