% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_add_names.R
\name{qp_add_names}
\alias{qp_add_names}
\alias{qp_add_names.list}
\alias{qp_add_names.data.frame}
\title{Add sample names}
\usage{
qp_add_names(x, ...)

\method{qp_add_names}{list}(x, sample_names = NULL, ...)

\method{qp_add_names}{data.frame}(x, sample_names = NULL, ...)
}
\arguments{
\item{x}{A \code{data.frame} (or a list containing one) that contains columns
\code{index} (which denotes sample number) and \code{sample_type}, which should be
either "unknown" or "standard".}

\item{...}{Unused}

\item{sample_names}{Optional character vector. If NULL, uses sample index.
In a standard workflow, the index is the order the sample appears in the
plate}
}
\description{
Add sample names
}
