% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_add_std_conc.R
\name{qp_add_std_conc}
\alias{qp_add_std_conc}
\alias{qp_add_std_conc.data.frame}
\alias{qp_add_std_conc.list}
\title{Add known concentrations of protein to standard samples}
\usage{
qp_add_std_conc(x, standard_scale = c(0, 2^((2:7) - 5)), ...)

\method{qp_add_std_conc}{data.frame}(x, standard_scale = c(0, 2^((2:7) - 5)), ...)

\method{qp_add_std_conc}{list}(x, standard_scale = c(0, 2^((2:7) - 5)), ...)
}
\arguments{
\item{x}{A \code{data.frame} containing a \code{sample_type} and \code{index} columns.
See details.}

\item{standard_scale}{A numeric vector giving the concentrations of the
standards. The units are arbitrary, but will determine the units of the
output concentrations.}
}
\value{
Same type as x, with a \code{.conc} column
}
\description{
Add known concentrations of protein to standard samples
}
\details{
Input is expected to have two columns:
\itemize{
\item \code{sample_type}: A character vector denoting which samples are standards
with "standard". All other values will be considered unknowns.
\item \code{index}: A numeric column denoting the sample number. Index 1 will
correspond to the first item in \code{standard_scale}, 2 will be the second,
etc.
}
}
