% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calc_abs_mean.R
\name{qp_calc_abs_mean}
\alias{qp_calc_abs_mean}
\title{Calculate absorbance means with optional outlier removal}
\usage{
qp_calc_abs_mean(
  x,
  exclude_outliers = c("all", "standards", "samples", "none")
)
}
\arguments{
\item{exclude_outliers}{Which sample types should have outliers excluded from
their mean calculations?}
}
\description{
Calculate absorbance means with optional outlier removal
}
