% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_calc_conc.R
\name{qp_calc_conc}
\alias{qp_calc_conc}
\title{Predict concentrations from standards fit}
\usage{
qp_calc_conc(x, ignore_outliers = TRUE, group_cols = c("sample_type", "index"))
}
\arguments{
\item{x}{A list. See details.}

\item{ignore_outliers}{Boolean. Should outliers be considered when
calculating the mean? See details.}

\item{group_cols}{Character vector. Columns to group by before taking the
mean.}
}
\value{
Returns a \code{list} with the input fit and data.frame, with additional
columns:
\itemize{
\item \code{.pred}: The predicted value from the provided model
\item \code{.pred_conc}: \code{.pred}, transformed by \code{conc_transform}
\item \code{.pred_conc_mean}: The mean of \code{.pred_conc}, sans samples where column
\code{.is_outlier == TRUE}
}
}
\description{
Predict concentrations from standards fit
}
\details{
The supplied list should contain two items - \code{fit}, generated by
\code{qp_fit}, and \code{qp}, a data.frame. \code{qp} should contain the following:
\itemize{
\item Columns used in \code{fit}. Usually, this is \code{.log2_abs}
\item Any columns in \code{group_cols}
\item If \code{ignore_outliers = TRUE}, \code{.is_outlier} will be used if supplied,
or created if not.
}
}
\examples{

data <- system.file("extdata", "absorbances.txt", package = "qp")
calculated <- qp(data, replicate_orientation = "h")

# Making a minimal object:
calculated$qp <- calculated$qp |>
  dplyr::select(
    .log2_abs, sample_type, index, .is_outlier
  )

calculated

qp_calc_conc(calculated)

}
