% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_mark_outliers.R
\name{qp_mark_outliers}
\alias{qp_mark_outliers}
\alias{qp_mark_outliers.data.frame}
\alias{qp_mark_outliers.list}
\title{Mark absorbance outliers}
\usage{
qp_mark_outliers(x, ignore_outliers = c("all", "standards", "samples", "none"))

\method{qp_mark_outliers}{data.frame}(x, ignore_outliers = c("all", "standards", "samples", "none"))

\method{qp_mark_outliers}{list}(x, ignore_outliers = c("all", "standards", "samples", "none"))
}
\arguments{
\item{x}{A \code{data.frame} or \code{list} containing a \code{data.frame} named \code{qp}. See
details.}

\item{ignore_outliers}{Which sample types should have outliers marked?}
}
\value{
The input \code{tibble} with an \code{.is_outlier} column
}
\description{
Mark absorbance outliers
}
\details{
Input \code{data.frame} must contain the following columns:
\itemize{
\item \code{sample_type}. Character. Must contain values either "standard" or
"unknown"
\item \code{index}. Numeric. Denotes sample number.
\item \code{.abs}. Numeric. Contains absorbance values.
}
}
