% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{qp_pal}
\alias{qp_pal}
\title{The default color palette for qp}
\format{
An object of class \code{character} of length 100.
}
\usage{
qp_pal
}
\description{
It attempts to match the real life colors of a protein quantification
experiment, in combination with \code{abs_to_col}
}
\keyword{datasets}
