% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_plot_plate.R
\name{qp_plot_plate}
\alias{qp_plot_plate}
\title{View the absorbances of an analyzed \code{qp} as they were on the plate}
\usage{
qp_plot_plate(x, size = 15)
}
\arguments{
\item{x}{A \code{data.frame} with \code{.row}, \code{.col}, and \code{.abs} columns}

\item{size}{The size of the points used to illustrate the wells. Passed to
geom_point.}
}
\value{
a \code{ggplot}
}
\description{
View the absorbances of an analyzed \code{qp} as they were on the plate
}
\examples{
qp_plot_plate(absorbances)
}
