% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_plot_standards.R
\name{qp_plot_standards}
\alias{qp_plot_standards}
\title{View an absorbance/concentration plot}
\usage{
qp_plot_standards(x)
}
\arguments{
\item{x}{The output of \code{qp()} (a \code{list}) or \code{qp_calc_conc()} (a \code{data.frame})}
}
\value{
a \code{ggplot}
}
\description{
View an absorbance/concentration plot
}
\examples{
absorbances |>
  qp() |>
  qp_plot_standards()
}
