% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_read.R
\name{qp_read}
\alias{qp_read}
\alias{qp_read.character}
\alias{qp_read.data.frame}
\alias{qp_read.gp}
\alias{qp_read.spectramax}
\title{Read in data into a common quantify protein format
This function will perform some very basic checks to see if the data is
properly formatted}
\usage{
qp_read(x, wavelength = 562, ...)

\method{qp_read}{character}(x, wavelength, ...)

\method{qp_read}{data.frame}(x, ...)

\method{qp_read}{gp}(x, ...)

\method{qp_read}{spectramax}(x, wavelength, ...)
}
\arguments{
\item{x}{A \code{gp}, \code{data.frame}/\code{tibble}, \code{spectramax}, or character
path to a raw SPECTRAmax .xls(x)/.txt}

\item{wavelength}{For SPECTRAmax files, the wavelength captured.}

\item{...}{Unused}
}
\value{
A \code{gp}
}
\description{
Read in data into a common quantify protein format
This function will perform some very basic checks to see if the data is
properly formatted
}
