% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_remove_empty.R
\name{qp_remove_empty}
\alias{qp_remove_empty}
\title{Remove empty wells from data}
\usage{
qp_remove_empty(x)
}
\arguments{
\item{x}{A \code{data.frame} containing columns \code{.pred_conc} and \code{sample_type}.
See details.}
}
\value{
Same as input
}
\description{
Remove empty wells from data
}
\details{
This function keeps any columns with positive \code{.pred_conc} or
\code{sample_type == "standard"}
}
