% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_report.R
\name{qp_report}
\alias{qp_report}
\title{Create a report for a protein quantificaiton experiment}
\usage{
qp_report(qp, output_file, other = list())
}
\arguments{
\item{qp}{Likely the output from \code{qp} AND \code{qp_dilute}.}

\item{output_file}{Character. The path of the file to export, including
\code{.html}}

\item{other}{Generally used for Shiny application. Assumes a named list of
key-values that will be used to document report parameters.}
}
\description{
Create a report for a protein quantificaiton experiment
}
\examples{
\dontrun{
absorbances |>
  qp() |>
  qp_dilute() |>
  qp_report(
    "~/my_report.html",
    other = list(key = "value") # Essentially metadata
  )
}
}
