% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_report.R
\name{qp_summarize}
\alias{qp_summarize}
\alias{qp_summarize.data.frame}
\alias{qp_summarize.list}
\title{Summarize output from qp pipeline}
\usage{
qp_summarize(x)

\method{qp_summarize}{data.frame}(x)

\method{qp_summarize}{list}(x)
}
\arguments{
\item{x}{A \code{data.frame} or a \code{list} containing a \code{data.frame} named \code{qp}}
}
\value{
A \code{tibble} with the sample \code{name}, \code{sample_type}, and the mean of
its predicted concentration (\code{.pred_conc_mean})
}
\description{
Summarize output from qp pipeline
}
