% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_tidy.R
\name{qp_tidy}
\alias{qp_tidy}
\title{Wrangle qp \code{gp} according to replicate orientation}
\usage{
qp_tidy(x, replicate_orientation, n_standards = 7, n_replicates = 3)
}
\arguments{
\item{x}{A \code{gp}}

\item{replicate_orientation}{Either 'h' or 'v' - see Details.}
}
\value{
a \code{tibble}
}
\description{
Wrangle qp \code{gp} according to replicate orientation
}
\details{
The standards must be in ascending concentration starting in the
upper left corner. Whether this is from from left to right or top to bottom
can be specified in 'replicate orientation'. Note that 'replicate
orientation' specified the direction that REPLICATES lie, NOT the direction
the samples flow (which will be opposite).
}
