% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_calc_abs_mean.R
\name{qp_calc_abs_mean}
\alias{qp_calc_abs_mean}
\title{Calculate absorbance means with optional outlier removal}
\usage{
qp_calc_abs_mean(x, ignore_outliers = c("all", "standards", "samples", "none"))
}
\arguments{
\item{x}{A \code{data.frame}. See details.}

\item{ignore_outliers}{Character. Which sample types should have outliers
excluded from their mean calculations?}
}
\description{
Calculate absorbance means with optional outlier removal
}
\details{
The supplied \code{data.frame} must include the following columns:
\itemize{
\item \code{sample_type}, which should contain only either "standard" or "unknown"
}
}
