% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_calc_abs_mean.R
\name{qp_calc_abs_mean}
\alias{qp_calc_abs_mean}
\alias{qp_calc_abs_mean.data.frame}
\alias{qp_calc_abs_mean.list}
\title{Calculate absorbance means with optional outlier removal}
\usage{
qp_calc_abs_mean(x, ignore_outliers = c("all", "standards", "samples", "none"))

\method{qp_calc_abs_mean}{data.frame}(x, ignore_outliers = c("all", "standards", "samples", "none"))

\method{qp_calc_abs_mean}{list}(x, ignore_outliers = c("all", "standards", "samples", "none"))
}
\arguments{
\item{x}{A \code{data.frame} or \code{list} containing a \code{data.frame} named \code{qp}. See
details.}

\item{ignore_outliers}{Which sample types should have outliers ignored from
their mean calculations? If \code{.is_outlier} column is supplied, this argument
is ignored.}
}
\value{
The input \code{tibble} with an \code{.is_outlier} column and a \code{.mean} column
}
\description{
Calculate absorbance means with optional outlier removal
}
\details{
Input \code{data.frame} must contain the following columns:
\itemize{
\item \code{sample_type}. Character. Must contain values either "standard" or
"unknown"
\item \code{index}. Numeric. Denotes sample number.
\item \code{.abs}. Numeric. Contains absorbance values.
\item If a boolean \code{.is_outlier} is supplied, that will be used instead.
}
}
\examples{
 library(dplyr)

abs <- expand.grid(
  sample_type = c("standard", "unknown"),
  index = 1:7,
  rep = 1:3
) |>
  dplyr::arrange(sample_type, index, rep)

abs$.abs <- abs(rnorm(nrow(abs), mean = abs$index))

# Selecting different subsets for outlier removal
qp_calc_abs_mean(abs, "none")

qp_calc_abs_mean(abs, "standards")

qp_calc_abs_mean(abs, "samples")

qp_calc_abs_mean(abs, "all")

# If an `.is_outlier` column is provided, that will be used instead:

abs$.is_outlier <- rep(c(TRUE, FALSE), length.out = nrow(abs))

qp_calc_abs_mean(abs)
}
