% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_calc_abs_mean.R
\name{qp_calc_abs_mean}
\alias{qp_calc_abs_mean}
\title{Calculate absorbance means with optional outlier removal}
\usage{
qp_calc_abs_mean(x, ignore_outliers = c("all", "standards", "samples", "none"))
}
\arguments{
\item{x}{A \code{data.frame}. See details.}

\item{ignore_outliers}{Which sample types should have outliers ignored from
their mean calculations?}
}
\value{
a \code{tibble} with an \code{.is_outlier} column and a \code{.mean} column
}
\description{
Calculate absorbance means with optional outlier removal
}
\details{
Input \code{data.frame} must contian the following columns:
\itemize{
\item \code{sample_type}. Character. Must contain values either "standard" or
"unknown"
\item \code{index}. Numeric. Denotes sample number.
\item \code{.abs}. Numeric. Contains absorbance values.
}
}
