% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_calc_conc.R
\name{qp_calc_conc}
\alias{qp_calc_conc}
\title{Predict concentrations from standards fit}
\usage{
qp_calc_conc(
  x,
  fit,
  ignore_outliers = TRUE,
  group_cols = c("sample_type", "index")
)
}
\arguments{
\item{x}{A data.frame. See details.}

\item{fit}{An \code{lm} object used to predict concentrations}

\item{group_cols}{Character vector. Columns to group by before
taking the mean.}

\item{ignore_outliers.}{Boolean. Should outliers be considered when
calculating the mean? See details.}
}
\value{
Returns a \code{tibble} with columns:
\itemize{
\item \code{.pred}: The predicted value from the provided model
\item \code{.pred_conc}: \code{.pred}, transformed by \code{conc_transform}
\item \code{.pred_conc_mean}: The mean of \code{.pred_conc}, sans samples where column
\code{.is_outlier == TRUE}
}
}
\description{
Predict concentrations from standards fit
}
\details{
The supplied data.frame should have the following columns:
\itemize{
\item Columns used in \code{fit}. Usually, this is \code{.log2_conc} and \code{.log2_abs}
\item Any columns in \code{group_cols}
\item If \code{ignore_outliers = TRUE}, \code{.is_outlier} will be used if supplied,
or created if not.
}
}
