% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_dilute.R
\name{qp_dilute}
\alias{qp_dilute}
\alias{qp_dilute.data.frame}
\alias{qp_dilute.list}
\title{Calculate dilutions for an analyzed \code{qp} \code{list}}
\usage{
qp_dilute(x, ...)

\method{qp_dilute}{data.frame}(x, target_conc, target_vol, ...)

\method{qp_dilute}{list}(x, target_conc, target_vol, ...)
}
\arguments{
\item{x}{A \code{data.frame} or \code{list} containing a \code{data.frame} named \code{qp} with
a column named \code{.pred_conc}}

\item{target_conc}{Numeric vector. Target concentration in (mg/mL) protein.
If length == 1, recycled.}

\item{target_vol}{Target volume in uL. If length == 1, recycled.}
}
\value{
A \code{tibble}, with the volumes of lysate and volumes of H2O to add.
}
\description{
Calculate dilutions for an analyzed \code{qp} \code{list}
}
