% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_plot_plate.R
\name{qp_plot_plate}
\alias{qp_plot_plate}
\title{View the absorbances of an analyzed \code{qp} as they were on the plate}
\usage{
qp_plot_plate(x)
}
\arguments{
\item{x}{A \code{data.frame} with \code{.row}, \code{.col}, and \code{.abs} columns}
}
\value{
a \code{ggplot}
}
\description{
View the absorbances of an analyzed \code{qp} as they were on the plate
}
\examples{
x <- expand.grid(.row = 1:8, .col = 1:12)
x$.abs <- 1:96
make_qp_plate_view(x)
}
