% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_tidy.R
\name{qp_tidy}
\alias{qp_tidy}
\alias{qp_tidy.character}
\alias{qp_tidy.spectramax}
\alias{qp_tidy.gp}
\alias{qp_tidy.default}
\title{Read in and wrangle protein quantification data}
\usage{
qp_tidy(x, ...)

\method{qp_tidy}{character}(x, ...)

\method{qp_tidy}{spectramax}(
  x,
  replicate_orientation = c("h", "v"),
  n_standards = 7,
  n_replicates = 3,
  wavelength = 562,
  ...
)

\method{qp_tidy}{gp}(x, ...)

\method{qp_tidy}{default}(x, ...)
}
\arguments{
\item{x}{A \code{gp}, \code{data.frame}/\code{tibble}, \code{spectramax}, or character path to a
raw SPECTRAmax .xls(x)/.txt}

\item{...}{Arguments passed to relevant methods.}

\item{replicate_orientation}{Character. Specified the direction the
\emph{replicates} lie, not the direction the samples flow (which will be
perpendicular to \code{replicate_orientation}).}

\item{n_standards}{Numeric. The number of different concentrations of
standards. Does not include replicates.}

\item{n_replicates}{Numeric. The number of replicates per sample.}

\item{wavelength}{Numeric. For SPECTRAmax files and objects, the wavelength
measured. Otherwise, ignored.}
}
\value{
a \code{data.frame}
}
\description{
Read in and wrangle protein quantification data
}
\details{
\code{qp} assumes that if you read in data not in a \code{spectramax} file or
object, you probably have a custom workflow in mind - therefore, tidying
will be minimal and mostly focused on checking for validity.
}
